refactor(run): split into components
